+++ src/fzero_mods.h
@@ -2,4 +2,4 @@
 #include "fzero_video.h"
 #include "recomp_launcher.h"
 const RecompLauncherCModProvider *FzeroModsProvider(FzeroVideoSettings *settings,
-                                                   const char *path);
+                                                   const char *path, bool bundled_music);